requirements.txt
flask
requests
python-dotenv
chromadb
sentence-transformers
